* feature/#7:
  #7 불필요한 주석 제거
  #7 taskEntity에서 courseId, week, day로 조회하는 api 추가
  #7 course에 속한 user목록, task 목록 조회

+++ src/main/java/com/project/feedback/domain/dto/task/TaskCreateRequest.java
@@ -16,8 +16,8 @@ public class TaskCreateRequest {
     private String description;
     private String status;
     private String courseName;
-    private Integer week;
-    private Integer day;
+    private Long week;
+    private Long day;
 
     public TaskEntity toEntity(User user, CourseEntity course) {
         return TaskEntity.builder()
