+++ src/main/java/com/project/feedback/domain/dto/task/TaskUpdateRequest.java
@@ -10,6 +10,6 @@ public class TaskUpdateRequest {
     private String title;
     private String description;
     private TaskStatus taskStatus;
-    private Integer week;
-    private Integer day;
+    private Long week;
+    private Long day;
 }
